renamed and refactored the receiver. Added Start method
